pyproject.toml
[tool.poetry]
name = "pdf-query"
version = "0.1.0"
description = ""
authors = ["hugobowne <hugobowne@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
gradio = "^4.44.0"
llama-index = "^0.11.9"
pymupdf = "^1.24.10"
llama-index-llms-ollama = "^0.3.2"
sentence-transformers = "^3.1.1"
llama-index-embeddings-huggingface = "^0.3.1"
# add burr[start,opentelemetry] to the dependencies
torch = "2.2.2"
burr = {extras = ["opentelemetry", "start"], version = "^0.32.0"}
opentelemetry-instrumentation-openai = "^0.33.3"
opentelemetry-instrumentation-ollama = "^0.33.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
